Fix messaging

+++ websitebackend/SocialMediaApp/src/main/java/SoloProject/SocialMediaApp/controller/PostController.java
@@ -1,21 +1,17 @@
 package SoloProject.SocialMediaApp.controller;
 
-import SoloProject.SocialMediaApp.models.Comment;
-import SoloProject.SocialMediaApp.models.CompressedImage;
-import SoloProject.SocialMediaApp.models.CreatePostRequest;
-import SoloProject.SocialMediaApp.models.Post;
-import SoloProject.SocialMediaApp.repository.AppUserRepository;
-import SoloProject.SocialMediaApp.repository.CommentRepository;
-import SoloProject.SocialMediaApp.repository.CompressedImageRepository;
-import SoloProject.SocialMediaApp.repository.PostRepository;
+import SoloProject.SocialMediaApp.models.*;
+import SoloProject.SocialMediaApp.repository.*;
+import SoloProject.SocialMediaApp.service.AccountService;
 import SoloProject.SocialMediaApp.service.CompressionService;
 import SoloProject.SocialMediaApp.service.PostService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.io.IOException;
 import java.sql.SQLException;
@@ -34,13 +30,21 @@ public class PostController {
     private final CompressedImageRepository compressedImageRepository;
     private final PostRepository postRepository;
 
+    private final NotificationRepository notificationRepository;
+
+    private final AccountService accountService;
+
+    private final AppUserRepository appUserRepository;
     @Autowired
-    public PostController(PostService postService, PostRepository postRepository, CompressionService compressionService, CompressedImageRepository compressedImageRepository, CommentRepository commentRepository) {
+    public PostController(PostService postService, NotificationRepository notificationRepository, PostRepository postRepository, CompressionService compressionService, CompressedImageRepository compressedImageRepository, CommentRepository commentRepository, AccountService accountService, AppUserRepository appUserRepository) {
         this.postService = postService;
         this.compressedImageRepository = compressedImageRepository;
         this.commentRepository = commentRepository;
         this.compressionService = compressionService;
         this.postRepository = postRepository;
+        this.accountService = accountService;
+        this.appUserRepository = appUserRepository;
+        this.notificationRepository = notificationRepository;
     }
 
     @GetMapping("/post/{postId}/comments")
@@ -67,11 +71,15 @@ public class PostController {
 
 
     @PostMapping("/post/{userId}/createPost")
-    public ResponseEntity<Post> createPost(
+    public ResponseEntity<?> createPost(
             @PathVariable Long userId,
-            @RequestBody CreatePostRequest createPostRequest // Assuming CreatePostRequest contains the Post and List<String> of base64 images
+            @RequestBody CreatePostRequest createPostRequest
     ) throws IOException, SQLException {
-        Post createdPost = postService.createPost(userId, createPostRequest.getPost());
+        Post createdPost = postService.createPost(userId, createPostRequest);
+
+        if (createdPost == null) {
+            return new ResponseEntity<>("Failed to create post", HttpStatus.BAD_REQUEST);
+        }
 
         if (createPostRequest.getImages() != null && !createPostRequest.getImages().isEmpty()) {
             postService.addImagesToPost(createdPost.getId(), createPostRequest.getImages());
@@ -88,22 +96,36 @@ public class PostController {
 
 
     @DeleteMapping("/post/{postId}/deletePost")
-    public ResponseEntity<?> deletePost(@PathVariable Long postId){
-        postService.deletePost(postId);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity<?> deletePost(@PathVariable Long postId, Authentication authentication) {
+        Optional<Post> post = postRepository.findById(postId);
+        if (post.isPresent()) {
+            if (!accountService.checkAuthenticationMatch(post.get().getPosterUsername(), authentication)) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access Denied");
+            }
+            postService.deletePost(postId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
 
     @PutMapping("/post/{postId}/{reactorId}/updateReactionPost")
-    public ResponseEntity<Post> updateReactionPost(
+    public ResponseEntity<?> updateReactionPost(
             @PathVariable Long reactorId,
             @PathVariable Long postId,
-            @RequestBody Map<String, String> payload) {
+            @RequestBody Map<String, String> payload,
+            Authentication authentication) {
+
+        AppUser appUser = appUserRepository.findByAppUserID(reactorId);
+        if (!accountService.checkAuthenticationMatch(appUser.getUsername(), authentication)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access Denied");
+        }
 
         String action = payload.get("action");
+        AppUser reactor = appUserRepository.findById(reactorId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Reactor not found"));
 
         Post updatedPost = postService.handlePostReaction(postId, reactorId, action);
-
         if (updatedPost != null) {
             return new ResponseEntity<>(updatedPost, HttpStatus.OK);
         } else {
@@ -113,9 +135,11 @@ public class PostController {
 
 
 
-    @GetMapping("/post/{userId}/friendPosts")
-    public ResponseEntity<List<Post>> getAllFriendPosts(@PathVariable Long userId) {
-        List<Post> posts = postService.getAllFriendPosts(userId);
+
+    @GetMapping("/post/{userId}/posts/{targetUserId}")
+    public ResponseEntity<List<Post>> getAllPostsByUserId(@PathVariable Long userId, @PathVariable Long targetUserId) {
+        List<Post> posts = postService.getPostsByPosterId(userId, targetUserId);
+
         if (posts != null && !posts.isEmpty()) {
             return new ResponseEntity<>(posts, HttpStatus.OK);
         } else {
@@ -123,15 +147,46 @@ public class PostController {
         }
     }
 
+    @GetMapping("/posts/{targetUserId}")
+    public ResponseEntity<List<Post>> getAllPublicPosts(@PathVariable Long targetUserId) {
+        List<Post> posts = postService.getAllUserPublicPosts(targetUserId);
 
-    @GetMapping("/post/{userId}/posts/{targetUserId}")
-    public ResponseEntity<List<Post>> getAllPostsByUserId(@PathVariable Long userId, @PathVariable Long targetUserId) {
-        List<Post> posts = postService.getPostsByPosterId(userId, targetUserId);
-        if (posts != null && !posts.isEmpty()) {
-            return new ResponseEntity<>(posts, HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if (posts == null) {
+            return new ResponseEntity<>(new ArrayList<>(), HttpStatus.OK);
         }
+        return new ResponseEntity<>(posts, HttpStatus.OK);
+    }
+
+    @GetMapping("/notifications/posts/{userId}/getPostsNotification")
+    public ResponseEntity<?> getUserPostNotificationCount(@PathVariable Long userId, Authentication authentication) {
+        System.out.println("Request userid: " + userId);
+        String username = appUserRepository.findByAppUserID(userId).getUsername();
+        if (!accountService.checkAuthenticationMatch(username, authentication)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access Denied");
+        }
+
+        return postService.getPostNotificationCount(userId);
+    }
+
+    @GetMapping("/notifications/posts/{userId}/{postId}/getIndividualPostNotification")
+    public ResponseEntity<?> getUserPostNotification(@PathVariable Long userId, @PathVariable Long postId, Authentication authentication){
+        System.out.println("Request userid: " + userId);
+        String username = appUserRepository.findByAppUserID(userId).getUsername();
+        if (!accountService.checkAuthenticationMatch(username, authentication)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access Denied");
+        }
+
+        return postService.getPostNotification(userId, postId);
+    }
+
+    @DeleteMapping("/notifications/posts/{userId}/delete/{postId}")
+    public ResponseEntity<?> deletePostNotification(@PathVariable Long userId, @PathVariable Long postId, Authentication authentication){
+        String username = appUserRepository.findByAppUserID(userId).getUsername();
+        if (!accountService.checkAuthenticationMatch(username, authentication)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access Denied");
+        }
+        notificationRepository.deleteByObjectIdAndUserIdAndType(postId, userId, "Tag");
+        return ResponseEntity.ok().build();
     }
